주관식1,2, 객관식1 문항 넣는 것까지 만듦. DB에 넣는 건 아직.

+++ src/main/java/com/freeneo/survey/mapper/SurveyMapper.java
@@ -7,7 +7,7 @@ import com.freeneo.survey.domain.Survey;
 public interface SurveyMapper {
 	public List<Survey> list();
 	public Survey select(Long id);
-	public int insert(Survey survey);
-	public int update(Survey survey);
-	public int delete(Long id);
+	public Long insert(Survey survey);
+	public Long update(Survey survey);
+	public Long delete(Long id);
 }
